+++ apps/ymir-sdl3/src/util/math.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <cmath>
+
+namespace util {
+
+double RoundToMultiple(double value, double multiple) {
+    return std::round(value / multiple) * multiple;
+}
+
+} // namespace util